REF: move atef.widgets.config.procedure into data_active
