+++ atef/widgets/__init__.py
@@ -1,3 +1,2 @@
 # TODO: Configuration GUI needs to be updated for ConfigurationGroup
 # from .config import *  # NOQA
-from .config.procedure import *  # NOQA
